+++ src/modules/server/l4/mod.rs
@@ -1,5 +1,6 @@
 /* src/modules/server/l4/mod.rs */
 
+pub mod dispatcher;
 pub mod fs;
 pub mod loader;
 pub mod model;